update component QB AnnotationOfSpotClassTGM for qanary 3.1.2 (https://github.com/anbo-de/team-tasks/issues/60) fix some TODOS
