+++ pom.xml
@@ -49,7 +49,7 @@
                 <module>qanary-component-QB-SimpleRealNameOfSuperHero</module>
                 <module>qanary-component-QB-ComicCharacterAlterEgoSimpleDBpedia</module>
                 <module>qa.qanary_component-QueryBuilder</module>
-                <module>qa.qanary_component-AnnotationofSpotClass-tgm</module>
+                <module>qanary-component-QB-AnnotationOfSpotClassTGM</module>
                 <module>qa.qanary_component-AnnotationofSpotProperty-tgm</module>
                 <module>qa.qanary_component-DiambiguationClass-OKBQA</module>
                 <module>qa.qanary_component-DiambiguationProperty-OKBQA</module>
