--- qa.qanary_component-AnnotationofSpotClass-tgm/src/main/java/eu/wdaqua/qanary/annotationofspotclass/Property.java
@@ -1,10 +1,0 @@
-package eu.wdaqua.qanary.annotationofspotclass;
-import java.util.List;
-import java.util.ArrayList;
-
-public class Property{
-	public List<String> property = new ArrayList<String>();
-	public List<String> resource = new ArrayList<String>();
-	public List<String> resourceL = new ArrayList<String>();
-	public List<String> classRdf = new ArrayList<String>();
-}
