+++ qanary-component-QB-AnnotationOfSpotClassTGM/Dockerfile
