+++ qanary-component-QB-AnnotationOfSpotClassTGM/README.md
@@ -1,0 +1,1 @@
+Component source: https://www.okbqa.org/architecture/template-generation
