+++ qanary-component-QB-AnnotationOfSpotClassTGM/pom.xml
@@ -4,18 +4,18 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>eu.wdaqua.qanary.component</groupId>
-    <artifactId>qa_qanary_component-AnnotationofSpotClass-tgm</artifactId>
-    <version>3.0.0</version>
+    <artifactId>qanary-component-QB-AnnotationOfSpotClassTGM</artifactId>
+    <version>3.1.0</version>
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.5.5</version>
+        <version>2.6.7</version>
     </parent>
     <properties>
         <java.version>11</java.version>
-        <qanary.version>[2.0.0,4.0.0)</qanary.version>
+        <qanary.version>[3.1.2,4.0.0)</qanary.version>
         <docker.image.prefix>qanary</docker.image.prefix>
-        <docker.image.name>annotation-of-spot-class</docker.image.name>
+        <docker.image.name>qanary-component-qb-annotationofspotclasstgm</docker.image.name>
         <dockerfile-maven-version>1.4.13</dockerfile-maven-version>
     </properties>
 
@@ -101,9 +101,19 @@
                             <goal>push</goal>
                         </goals>
                     </execution>
+                    <execution>
+                        <id>default-2</id>
+                        <goals>
+                            <goal>tag</goal>
+                            <goal>push</goal>
+                        </goals>
+                        <configuration>
+                            <tag>latest</tag>
+                        </configuration>
+                    </execution>
                 </executions>
                 <configuration>
-                    <repository>${docker.image.name}</repository>
+                    <repository>${docker.image.prefix}/${docker.image.name}</repository>
                     <tag>${project.version}</tag>
                     <buildArgs>
                         <JAR_FILE>${project.build.finalName}.jar</JAR_FILE>
