+++ qanary-component-QB-AnnotationOfSpotClassTGM/src/main/java/eu/wdaqua/qanary/annotationofspotclass/MySelection.java
@@ -1,8 +1,9 @@
 package eu.wdaqua.qanary.annotationofspotclass;
+
 public class MySelection {
-	public String word= "";
-	public String type="";
-	public String rsc="";
+    public String word = "";
+    public String type = "";
+    public String rsc = "";
     public int begin;
     public int end;
 }
