+++ qanary-component-QB-AnnotationOfSpotClassTGM/src/main/java/eu/wdaqua/qanary/annotationofspotclass/Property.java
@@ -1,0 +1,11 @@
+package eu.wdaqua.qanary.annotationofspotclass;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Property {
+    public List<String> property = new ArrayList<String>();
+    public List<String> resource = new ArrayList<String>();
+    public List<String> resourceL = new ArrayList<String>();
+    public List<String> classRdf = new ArrayList<String>();
+}
