+++ qanary-component-QB-AnnotationOfSpotClassTGM/src/main/resources/config/application.properties
@@ -24,6 +24,8 @@ logging.level.org.springframework.jmx:WARN
 logging.level.org.springframework.mock:WARN
 logging.level.org.springframework.test:WARN
 logging.level.eu.wdaqua.qanary:DEBUG
+# show colors in IDE console
+spring.output.ansi.enabled=always
 ### SSL configuration
 # the path to the key store that contains the SSL certificate, e.g., classpath:keystore.p12
 server.ssl.key-store=
@@ -33,4 +35,5 @@ server.ssl.key-store-password=
 server.ssl.key-store-type=
 # toggle whether HTTP or HTTPS should be used (if SSL is set up)
 server.ssl.enabled=false
-
+annotationOfSpotClass.url=http://ws.okbqa.org:1515/templategeneration/rocknrole
+#annotationOfSpotClass.url=http://121.254.173.90:1515/templategeneration/rocknrole
